Updated Editor

+++ BahnEditor/BahnEditor.BahnLib/Constants.cs
@@ -9,13 +9,13 @@
 
 		//Layers
 		public const uint GSYMTEIL_MAXZAHL = 4;	//Maximale Anzahl von Layern pro Element
-		public const uint GFX_Z_HI = 1;			//zum Hintergrund (=flach, nicht "aufwärts", z. B. Platformen)
-		public const uint GFY_Z_VG = 2;			//Vordergrund (in Front von Fahrwegen und Zügen
-		public const uint GFX_Z_HG = 3;			//Hintergrund (auch hinter Zügen auf ny+1)
-		public const uint GFX_Z_OG = 4;			//Front
-		public const uint GFX_Z_VGO = 5;		//Upper foreground
-		public const uint GFX_Z_MIN = 1;		//Minimum Layers
-		public const uint GFX_Z_MAX = 5;		//Maximum Layers
+		public const short GFX_Z_HI = 1;			//zum Hintergrund (=flach, nicht "aufwärts", z. B. Platformen)
+		public const short GFY_Z_VG = 2;			//Vordergrund (in Front von Fahrwegen und Zügen
+		public const short GFX_Z_HG = 3;			//Hintergrund (auch hinter Zügen auf ny+1)
+		public const short GFX_Z_OG = 4;			//Front
+		public const short GFX_Z_VGO = 5;		//Upper foreground
+		public const short GFX_Z_MIN = 1;		//Minimum Layers
+		public const short GFX_Z_MAX = 5;		//Maximum Layers
 
 		//Variants
 		public const uint KEINE_ALTERNATIVE = 0;
